graph: Move BlockHash into blockchain::types
